feat: confirm dialog before remove a book

+++ src/app/components/book-shelf/book-shelf.component.ts
@@ -61,6 +61,7 @@ export class BookShelfComponent {
   }
 
   removeBook(id: string) {
+    if (!confirm('Are you sure you want to delete this book?')) return;
     this.bookService.removeBook(id);
   }
 }
